fix: 캐로셀 로직을 공통 hook으로 수정
